inv-management-app/src/main/model/Event.java
package model;

import java.util.Calendar;
import java.util.Date;

// SOURCE: methods in this class modelled after Event Class from AlarmSystem
//         https://github.students.cs.ubc.ca/CPSC210/AlarmSystem

// represents a store inventory event
public class Event {
    private static final int HASH_CONSTANT = 13;
    private Date dateLogged;
    private String description;

    // EFFECTS: creates an event with the given description including
    //          the date and time the event was created
    public Event(String description) {
        dateLogged = Calendar.getInstance().getTime();
        this.description = description;
    }

    // EFFECTS: return date of event
    public Date getDate() {
        return dateLogged;
    }

    // EFFECTS: return description of event
    public String getDescription() {
        return description;
    }

    @Override
    public boolean equals(Object other) {
        if (other == null) {
            return false;
        }

        if (other.getClass() != this.getClass()) {
            return false;
        }

        Event otherEvent = (Event) other;

        return (this.dateLogged.equals(otherEvent.dateLogged)
                && this.description.equals(otherEvent.description));
    }

    @Override
    public int hashCode() {
        return (HASH_CONSTANT * dateLogged.hashCode() + description.hashCode());
    }

    @Override
    public String toString() {
        return dateLogged.toString() + "\n" + description;
    }

}
